+++ src/main/java/com/samuelcastro/ProyectoFinal/controllers/DepartamentoController.java
@@ -2,6 +2,7 @@ package com.samuelcastro.ProyectoFinal.controllers;
 
 import com.samuelcastro.ProyectoFinal.entities.Departamento;
 import com.samuelcastro.ProyectoFinal.services.DepartamentoService;
+import com.samuelcastro.ProyectoFinal.services.RegistroService;
 import com.samuelcastro.ProyectoFinal.services.ProfesorDetails;
 import com.samuelcastro.ProyectoFinal.utils.SecurityUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,6 +19,9 @@ public class DepartamentoController {
     @Autowired
     private DepartamentoService departamentoService;
 
+    @Autowired
+    private RegistroService registroService;
+
     @GetMapping("/nuevo")
     public String mostrarFormularioNuevoDepartamento(Model model) {
         model.addAttribute("departamento", new Departamento());
@@ -28,6 +32,8 @@ public class DepartamentoController {
     @PostMapping
     public String crearDepartamento(Departamento departamento) {
         departamentoService.save(departamento);
+        ProfesorDetails profesorDetails = SecurityUtils.getAuthenticatedUser();
+        registroService.registrarOperacion("Departamento", departamento.getIdDepartamento(), profesorDetails.getProfesor().getNombre() + " " + profesorDetails.getProfesor().getApellidos() + " EJECUTA CREAR ", departamento.getNombre());
         return "redirect:/api/departamentos";
     }
 
@@ -50,12 +56,19 @@ public class DepartamentoController {
     @PostMapping("/update")
     public String actualizarDepartamento(Departamento departamento) {
         departamentoService.save(departamento);
+        ProfesorDetails profesorDetails = SecurityUtils.getAuthenticatedUser();
+        registroService.registrarOperacion("Departamento", departamento.getIdDepartamento(), profesorDetails.getProfesor().getNombre() + " " + profesorDetails.getProfesor().getApellidos() + " EJECUTA ACTUALIZAR ", departamento.getNombre());
         return "redirect:/api/departamentos";
     }
 
     @GetMapping("/delete/{id}")
     public String borrarDepartamento(@PathVariable int id) {
-        departamentoService.deleteById(id);
+        Departamento departamento = departamentoService.findById(id);
+        if (departamento != null) {
+            departamentoService.deleteById(id);
+            ProfesorDetails profesorDetails = SecurityUtils.getAuthenticatedUser();
+            registroService.registrarOperacion("Departamento", id, profesorDetails.getProfesor().getNombre() + " " + profesorDetails.getProfesor().getApellidos() + " EJECUTA ELIMINAR ", departamento.getNombre());
+        }
         return "redirect:/api/departamentos";
     }
