+++ src/main/java/com/samuelcastro/ProyectoFinal/controllers/PrestamoController.java
@@ -4,6 +4,7 @@ import com.samuelcastro.ProyectoFinal.entities.Prestamo;
 import com.samuelcastro.ProyectoFinal.services.AlumnoService;
 import com.samuelcastro.ProyectoFinal.services.LibroService;
 import com.samuelcastro.ProyectoFinal.services.PrestamoService;
+import com.samuelcastro.ProyectoFinal.services.RegistroService;
 import com.samuelcastro.ProyectoFinal.services.ProfesorDetails;
 import com.samuelcastro.ProyectoFinal.utils.SecurityUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,6 +27,9 @@ public class PrestamoController {
 
     @Autowired
     private LibroService libroService;
+
+    @Autowired
+    private RegistroService registroService;
 
     @GetMapping
     public String getAllPrestamos(Model model) {
@@ -57,13 +61,47 @@ public class PrestamoController {
             prestamo.setProfesor(profesorDetails.getProfesor());
         }
         prestamo.setFechaPrestamo(new Date());
+        registroService.registrarOperacion("Prestamo", prestamo.getIdPrestamo(), prestamo.getProfesor().getNombre() + " " + prestamo.getProfesor().getApellidos() + " EJECUTA CREAR SOBRE ", libroService.findById(prestamo.getLibro().getIdLibro()).getTitulo());
         prestamoService.save(prestamo);
+        return "redirect:/api/prestamos";
+    }
+
+    @GetMapping("/editar/{id}")
+    public String mostrarFormularioEditarPrestamo(@PathVariable int id, Model model) {
+        Prestamo prestamo = prestamoService.findById(id);
+        model.addAttribute("prestamo", prestamo);
+        model.addAttribute("alumnos", alumnoService.findAll());
+        model.addAttribute("libros", libroService.findAll());
+        ProfesorDetails profesorDetails = SecurityUtils.getAuthenticatedUser();
+        if (profesorDetails != null) {
+            model.addAttribute("profesor", profesorDetails.getProfesor());
+        }
+        return "prestamos/editar-prestamo";
+    }
+
+    @PostMapping("/{id}")
+    public String actualizarPrestamo(@PathVariable int id, @ModelAttribute Prestamo prestamo) {
+        Prestamo existingPrestamo = prestamoService.findById(id);
+        if (existingPrestamo != null) {
+            existingPrestamo.setFechaPrestamo(prestamo.getFechaPrestamo());
+            existingPrestamo.setFechaPlazo(prestamo.getFechaPlazo());
+            existingPrestamo.setFechaDevolucion(prestamo.getFechaDevolucion());
+            existingPrestamo.setDevuelto(prestamo.isDevuelto());
+            prestamoService.save(existingPrestamo);
+            ProfesorDetails profesorDetails = SecurityUtils.getAuthenticatedUser();
+            registroService.registrarOperacion("Prestamo", existingPrestamo.getIdPrestamo(), profesorDetails.getProfesor().getNombre() + " " + profesorDetails.getProfesor().getApellidos() + " EJECUTA ACTUALIZAR SOBRE ", libroService.findById(existingPrestamo.getLibro().getIdLibro()).getTitulo());
+        }
         return "redirect:/api/prestamos";
     }
 
     @GetMapping("/eliminar/{id}")
     public String eliminarPrestamo(@PathVariable int id) {
-        prestamoService.deleteById(id);
+        Prestamo prestamo = prestamoService.findById(id);
+        if (prestamo != null) {
+            prestamoService.deleteById(id);
+            ProfesorDetails profesorDetails = SecurityUtils.getAuthenticatedUser();
+            registroService.registrarOperacion("Prestamo", id, profesorDetails.getProfesor().getNombre() + " " + profesorDetails.getProfesor().getApellidos() + " EJECUTA ELIMINAR SOBRE ", libroService.findById(prestamo.getLibro().getIdLibro()).getTitulo());
+        }
         return "redirect:/api/prestamos";
     }
 
@@ -74,6 +112,8 @@ public class PrestamoController {
             prestamo.setDevuelto(true);
             prestamo.setFechaDevolucion(new Date());
             prestamoService.save(prestamo);
+            ProfesorDetails profesorDetails = SecurityUtils.getAuthenticatedUser();
+            registroService.registrarOperacion("Prestamo", prestamo.getIdPrestamo(), profesorDetails.getProfesor().getNombre() + " " + profesorDetails.getProfesor().getApellidos() + " EJECUTA DEVOLVER SOBRE ", libroService.findById(prestamo.getLibro().getIdLibro()).getTitulo());
         }
         return "redirect:/api/prestamos";
     }
